Rename dissolved oxygen util file to fix typo
